--- packages/app/src/kits/Buttons/ButtonSubmit.tsx
@@ -1,81 +1,0 @@
-// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { motion } from 'framer-motion';
-import type { ComponentBaseWithClassName } from '@w3ux/types';
-import type { ButtonCommonProps, ButtonIconProps } from './types';
-import { onMouseHandlers } from './Utils';
-import { appendOr, appendOrEmpty } from '@w3ux/utils';
-
-export type ButtonSubmitProps = ComponentBaseWithClassName &
-  ButtonIconProps &
-  ButtonCommonProps & {
-    // use secondary network color.
-    colorSecondary?: boolean;
-    // button text.
-    text: string;
-    // large button, small otherwise.
-    lg?: boolean;
-    // pulsing effect.
-    pulse?: boolean;
-  };
-
-/**
- * @name ButtonSubmit
- * @description Submit button style used within modals to submit transactions.
- */
-export const ButtonSubmit = ({
-  colorSecondary,
-  disabled,
-  grow,
-  iconLeft,
-  iconRight,
-  iconTransform,
-  marginLeft,
-  marginRight,
-  marginX,
-  className,
-  style,
-  text,
-  lg,
-  pulse,
-  onClick,
-  onMouseOver,
-  onMouseMove,
-  onMouseOut,
-}: ButtonSubmitProps) => (
-  <motion.button
-    whileHover={{ scale: !disabled ? 1.02 : 1 }}
-    whileTap={{ scale: !disabled ? 0.98 : 1 }}
-    className={`btn-submit${appendOr(lg, 'lg', 'sm')}${appendOrEmpty(
-      colorSecondary,
-      'secondary-color'
-    )}${appendOrEmpty(grow, 'grow')}${appendOrEmpty(marginRight, 'm-right')}${appendOrEmpty(
-      marginLeft,
-      'm-left'
-    )}${appendOrEmpty(marginX, 'm-x')}${appendOrEmpty(pulse, 'pulse')}${
-      className ? ` ${className}` : ''
-    }`}
-    style={style}
-    type="button"
-    disabled={disabled}
-    {...onMouseHandlers({ onClick, onMouseOver, onMouseMove, onMouseOut })}
-  >
-    {iconLeft ? (
-      <FontAwesomeIcon
-        icon={iconLeft}
-        className={text ? 'icon-left' : undefined}
-        transform={iconTransform ? iconTransform : undefined}
-      />
-    ) : null}
-    {text ? text : null}
-    {iconRight ? (
-      <FontAwesomeIcon
-        icon={iconRight}
-        className={text ? 'icon-right' : undefined}
-        transform={iconTransform ? iconTransform : undefined}
-      />
-    ) : null}
-  </motion.button>
-);
