+++ packages/app/src/kits/Buttons/index.scss
@@ -119,26 +119,6 @@
 
 /* Button Styles */
 
-.btn-submit {
-  @include btn-core;
-  @include btn-layout;
-  @include btn-spacing;
-  @include btn-icons;
-  @include btn-submit;
-  @include btn-disabled;
-  @include btn-pulse;
-
-  border: 1px solid var(--accent-color-primary);
-  border-radius: var(--button-border-radius-small);
-  background: var(--accent-color-primary);
-  color: white;
-
-  &.secondary-color {
-    border-color: var(--accent-color-secondary);
-    background: var(--accent-color-secondary);
-  }
-}
-
 .btn-submit-invert {
   @include btn-core;
   @include btn-layout;
