+++ packages/app/src/library/SubmitTx/ManualSign/Vault/index.tsx
@@ -10,7 +10,7 @@ import { EstimatedTxFee } from 'library/EstimatedTxFee';
 import { useImportedAccounts } from 'contexts/Connect/ImportedAccounts';
 import type { SubmitProps } from '../../types';
 import { SignPrompt } from './SignPrompt';
-import { ButtonSubmit } from 'kits/Buttons/ButtonSubmit';
+import { ButtonSubmit } from 'ui-buttons';
 import { ButtonSubmitLarge } from 'library/SubmitTx/ButtonSubmitLarge';
 import { appendOrEmpty } from '@w3ux/utils';
