+++ packages/app/src/modals/DismissTips/index.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { usePlugins } from 'contexts/Plugins';
 import { Title } from 'library/Modal/Title';
 import { useOverlay } from 'kits/Overlay/Provider';
-import { ButtonSubmit } from 'kits/Buttons/ButtonSubmit';
+import { ButtonSubmit } from 'ui-buttons';
 import { ModalPadding } from 'kits/Overlay/structure/ModalPadding';
 
 export const DismissTips = () => {
