+++ packages/ui-buttons/src/ButtonSubmit/index.module.scss
@@ -1,0 +1,49 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.btnSubmit {
+  // @include btn-pulse;
+
+  border: 1px solid var(--accent-color-primary);
+  border-radius: var(--button-border-radius-small);
+  background: var(--accent-color-primary);
+  color: white;
+}
+
+.btnSubmit_secondaryColor {
+  border-color: var(--accent-color-secondary);
+  background: var(--accent-color-secondary);
+}
+
+.btnSubmit_sm {
+  padding: var(--button-padding-large);
+  font-size: var(--button-font-size-small);
+}
+
+.btnSubmit_lg {
+  padding: var(--button-padding-large);
+  font-size: var(--button-font-size-large);
+}
+
+.btnSubmit_pulse {
+  box-shadow: 0 0 30px 0 var(--accent-color-pending);
+  transform: scale(1);
+  animation: pulse 2s infinite;
+
+  @keyframes pulse {
+    0% {
+      transform: scale(0.95);
+      box-shadow: 0 0 0 0 var(--accent-color-pending);
+    }
+
+    70% {
+      transform: scale(1);
+      box-shadow: 0 0 0 10px rgb(0 0 0 / 0%);
+    }
+
+    100% {
+      transform: scale(0.95);
+      box-shadow: 0 0 0 0 rgb(0 0 0 / 0%);
+    }
+  }
+}
