+++ packages/ui-buttons/src/index.tsx
@@ -8,5 +8,6 @@ export * from './ButtonOption';
 export * from './ButtonPrimary';
 export * from './ButtonPrimaryInvert';
 export * from './ButtonSecondary';
+export * from './ButtonSubmit';
 
 export * from './ButtonText';
